+++ cmd/operator-sdk/run/cmd.go
@@ -33,7 +33,9 @@ import (
 type runCmd struct {
 	// Common options.
 	kubeconfig string
-	namespace  string
+	//TODO: remove namespace flag before 1.0.0
+	//namespace is deprecated
+	namespace string
 
 	// Run type.
 	olm, local bool
@@ -65,7 +67,15 @@ func NewCmd() *cobra.Command {
 			switch {
 			case c.olm:
 				c.olmArgs.KubeconfigPath = c.kubeconfig
-				c.olmArgs.OperatorNamespace = c.namespace
+				// operator-namespace flag is not set
+				// use default namespace from kubeconfig to deploy operator resources
+				if !cmd.Flags().Changed("operator-namespace") {
+					_, defaultNamespace, err := k8sinternal.GetKubeconfigAndNamespace(c.kubeconfig)
+					if err != nil {
+						return fmt.Errorf("error getting kubeconfig and default namespace: %v", err)
+					}
+					c.olmArgs.OperatorNamespace = defaultNamespace
+				}
 				if c.olmArgs.ManifestsDir == "" {
 					operatorName := filepath.Base(projutil.MustGetwd())
 					c.olmArgs.ManifestsDir = filepath.Join(olmcatalog.OLMCatalogDir, operatorName)
@@ -74,16 +84,25 @@ func NewCmd() *cobra.Command {
 					log.Fatalf("Failed to run operator using OLM: %v", err)
 				}
 			case c.local:
+				//TODO: remove namespace flag before 1.0.0
+				// set --watch-namespace flag if the --namespace flag is set
+				// (only if --watch-namespace flag is not set)
+				if cmd.Flags().Changed("namespace") {
+					log.Info("--namespace is deprecated; use --watch-namespace instead.")
+					if !cmd.Flags().Changed("watch-namespace") {
+						err := cmd.Flags().Set("watch-namespace", c.namespace)
+						return err
+					}
+				}
 				// Get default namespace to watch if unset.
-				if !cmd.Flags().Changed("namespace") {
+				if !cmd.Flags().Changed("watch-namespace") {
 					_, defaultNamespace, err := k8sinternal.GetKubeconfigAndNamespace(c.kubeconfig)
 					if err != nil {
 						return fmt.Errorf("error getting kubeconfig and default namespace: %v", err)
 					}
-					c.namespace = defaultNamespace
+					c.localArgs.watchNamespace = defaultNamespace
 				}
 				c.localArgs.kubeconfig = c.kubeconfig
-				c.localArgs.namespace = c.namespace
 				if err := c.localArgs.run(); err != nil {
 					log.Fatalf("Failed to run operator locally: %v", err)
 				}
@@ -98,9 +117,11 @@ func NewCmd() *cobra.Command {
 	cmd.Flags().StringVar(&c.kubeconfig, "kubeconfig", "",
 		"The file path to kubernetes configuration file. Defaults to location "+
 			"specified by $KUBECONFIG, or to default file rules if not set")
+	// Deprecated: namespace exists for historical compatibility. Use watch-namespace instead.
+	//TODO: remove namespace flag before 1.0.0
 	cmd.Flags().StringVar(&c.namespace, "namespace", "",
-		"The namespace where the operator watches for changes.")
-
+		"(Deprecated: use --watch-namespace instead.)"+
+			"The namespace where the operator watches for changes.")
 	// 'run --olm' and related flags.
 	cmd.Flags().BoolVar(&c.olm, "olm", false,
 		"The operator to be run will be managed by OLM in a cluster. "+