+++ cmd/operator-sdk/run/local.go
@@ -41,7 +41,7 @@ import (
 
 type runLocalArgs struct {
 	kubeconfig           string
-	namespace            string
+	watchNamespace       string
 	operatorFlags        string
 	ldFlags              string
 	enableDelve          bool
@@ -51,6 +51,10 @@ type runLocalArgs struct {
 
 func (c *runLocalArgs) addToFlags(fs *pflag.FlagSet) {
 	prefix := "[local only] "
+
+	fs.StringVar(&c.watchNamespace, "watch-namespace", "",
+		prefix+"The namespace where the operator watches for changes. Set \"\" for AllNamespaces, "+
+			"set \"ns1,ns2\" for MultiNamespace")
 	fs.StringVar(&c.operatorFlags, "operator-flags", "",
 		prefix+"The flags that the operator needs. Example: \"--flag1 value1 --flag2=value2\"")
 	fs.StringVar(&c.ldFlags, "go-ldflags", "", prefix+"Set Go linker options")
@@ -59,7 +63,7 @@ func (c *runLocalArgs) addToFlags(fs *pflag.FlagSet) {
 }
 
 func (c runLocalArgs) run() error {
-	log.Infof("Running the operator locally in namespace %s.", c.namespace)
+	log.Infof("Running the operator locally in namespace %s.", c.watchNamespace)
 
 	switch t := projutil.GetOperatorType(); t {
 	case projutil.OperatorTypeGo:
@@ -118,7 +122,7 @@ func (c runLocalArgs) runGo() error {
 	if c.kubeconfig != "" {
 		dc.Env = append(dc.Env, fmt.Sprintf("%v=%v", k8sutil.KubeConfigEnvVar, c.kubeconfig))
 	}
-	dc.Env = append(dc.Env, fmt.Sprintf("%v=%v", k8sutil.WatchNamespaceEnvVar, c.namespace))
+	dc.Env = append(dc.Env, fmt.Sprintf("%v=%v", k8sutil.WatchNamespaceEnvVar, c.watchNamespace))
 
 	// Set the ANSIBLE_ROLES_PATH
 	if c.ansibleOperatorFlags != nil && len(c.ansibleOperatorFlags.AnsibleRolesPath) > 0 {
@@ -136,7 +140,7 @@ func (c runLocalArgs) runGo() error {
 
 func (c runLocalArgs) runAnsible() error {
 	logf.SetLogger(zap.Logger())
-	if err := setupOperatorEnv(c.kubeconfig, c.namespace); err != nil {
+	if err := setupOperatorEnv(c.kubeconfig, c.watchNamespace); err != nil {
 		return err
 	}
 	return ansible.Run(c.ansibleOperatorFlags)
@@ -144,7 +148,7 @@ func (c runLocalArgs) runAnsible() error {
 
 func (c runLocalArgs) runHelm() error {
 	logf.SetLogger(zap.Logger())
-	if err := setupOperatorEnv(c.kubeconfig, c.namespace); err != nil {
+	if err := setupOperatorEnv(c.kubeconfig, c.watchNamespace); err != nil {
 		return err
 	}
 	return helm.Run(c.helmOperatorFlags)